+++ lib/recipes/dev-math/recipe.rb
@@ -136,5 +136,6 @@ DEV_MATH_RECIPE = Kitchen::BookRecipe.new(book_short_name: :dev_math) do |doc, _
   BakeToc.v1(book: book)
   BakeLinkPlaceholders.v1(book: book)
   BakeFolio.v1(book: book)
+  BakeRexWrappers.v1(book: book)
   BakeLinks.v1(book: book)
 end