+++ lib/recipes/nursing-internal/recipe.rb
@@ -164,5 +164,6 @@ do |doc, _resources|
   BakeToc.v1(book: book)
   BakeLinkPlaceholders.v1(book: book)
   BakeFolio.v1(book: book)
+  BakeRexWrappers.v1(book: book)
   BakeLinks.v1(book: book)
 end