+++ lib/recipes/statistics/recipe.rb
@@ -153,5 +153,6 @@ STATISTICS_RECIPE = Kitchen::BookRecipe.new(book_short_name: :statistics) do |do
   BakeLinkPlaceholders.v1(book: book)
   BakeToc.v1(book: book)
   BakeFolio.v1(book: book)
+  BakeRexWrappers.v1(book: book)
   BakeLinks.v1(book: book)
 end