+++ lib/recipes/world-history/recipe.rb
@@ -134,5 +134,6 @@ WORLD_HISTORY_RECIPE = Kitchen::BookRecipe.new(book_short_name: :world_history) 
   BakeLinkPlaceholders.v1(book: book)
   BakeToc.v1(book: book)
   BakeFolio.v1(book: book)
+  BakeRexWrappers.v1(book: book)
   BakeLinks.v1(book: book)
 end