+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+* Add section titles links to `summary` in `nursing-internal`
 * Change EOC titles in `nursing-internal`
 
 ## [v2.5.0] - 2023-11-03