+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+
+* Add `neurscience across species` note in `neuroscience`
 * Bake `learning-objectives` in `neuroscience`
 * Add `BakeRexWrappers` to every recipe-- gives an attribute for rex
 * Fix problem letter tag in multipart questions