+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [v2.6.0] - 2023-11-17
+
 * Add `boxed-feature` note in `neuroscience`
 * Bake EOC sections in  `neuroscience`
 * Add `neurscience across species` note in `neuroscience`